refactor(editor): improve implementation of lit adapter (#12101)
