+++ blocksuite/affine/widgets/slash-menu/src/view.ts
@@ -16,6 +16,7 @@ export class SlashMenuViewExtension extends ViewExtensionProvider {
 
   override setup(context: ViewExtensionContext) {
     super.setup(context);
+    if (this.isMobile(context.scope)) return;
     context.register(SlashMenuExtension);
   }
 }
