+++ packages/frontend/component/src/lit-react/lit-portal/index.ts
@@ -1,5 +1,6 @@
 export {
   type ElementOrFactory,
+  type ReactToLit,
   useLitPortal,
   useLitPortalFactory,
 } from './lit-portal';
