--- packages/frontend/core/src/blocksuite/extensions/index.ts
@@ -1,1 +1,0 @@
-export * from './entry/enable-mobile';
