+++ packages/frontend/core/src/blocksuite/extensions/reference-renderer.ts
@@ -1,20 +1,19 @@
-import type { ElementOrFactory } from '@affine/component';
+import type { ReactToLit } from '@affine/component';
 import type { AffineReference } from '@blocksuite/affine/inlines/reference';
 import { ReferenceNodeConfigExtension } from '@blocksuite/affine/inlines/reference';
 import type { ExtensionType } from '@blocksuite/affine/store';
-import type { TemplateResult } from 'lit';
 
 export type ReferenceReactRenderer = (
   reference: AffineReference
 ) => React.ReactElement;
 
 export function patchReferenceRenderer(
-  reactToLit: (element: ElementOrFactory) => TemplateResult,
+  reactToLit: ReactToLit,
   reactRenderer: ReferenceReactRenderer
 ): ExtensionType {
   const customContent = (reference: AffineReference) => {
     const node = reactRenderer(reference);
-    return reactToLit(node);
+    return reactToLit(node, true);
   };
 
   return ReferenceNodeConfigExtension({
