+++ packages/frontend/core/src/modules/doc-info/use-std.ts
@@ -1,0 +1,20 @@
+import { getViewManager } from '@affine/core/blocksuite/manager/migrating-view';
+import { DebugLogger } from '@affine/debug';
+import { BlockStdScope } from '@blocksuite/affine/std';
+import type { Store } from '@blocksuite/affine/store';
+import { useMemo } from 'react';
+
+const logger = new DebugLogger('doc-info');
+// todo(pengx17): use rc pool?
+export function createBlockStdScope(doc: Store) {
+  logger.debug('createBlockStdScope', doc.id);
+  const std = new BlockStdScope({
+    store: doc,
+    extensions: getViewManager().get('page'),
+  });
+  return std;
+}
+
+export function useBlockStdScope(doc: Store) {
+  return useMemo(() => createBlockStdScope(doc), [doc]);
+}
