+++ packages/frontend/core/src/modules/doc-info/views/database-properties/cells/rich-text.tsx
@@ -10,7 +10,7 @@ import { type CSSProperties, useEffect, useRef, useState } from 'react';
 import type * as Y from 'yjs';
 
 import type { DatabaseCellRendererProps } from '../../../types';
-import { useBlockStdScope } from '../../../utils';
+import { useBlockStdScope } from '../../../use-std';
 import * as styles from './rich-text.css';
 
 // todo(@pengx17): handle markdown/keyboard shortcuts
